re-organised nodes setup
- export `DEFAULT_NODES` as part of the library
- renamed `aside.js` to `AsideNode.js` to match lexical naming
