--- packages/koenig-lexical-playground/src/lib/nodes/DefaultNodes.js
@@ -1,19 +1,0 @@
-import {ListItemNode, ListNode} from '@lexical/list';
-import {CodeNode} from '@lexical/code';
-import {HeadingNode, QuoteNode} from '@lexical/rich-text';
-import {LinkNode} from '@lexical/link';
-import {AsideNode} from './aside';
-import {ImageNode} from './ImageNode';
-
-const DefaultNodes = [
-    HeadingNode,
-    ListNode,
-    ListItemNode,
-    QuoteNode,
-    AsideNode,
-    CodeNode,
-    LinkNode,
-    ImageNode
-];
-
-export default DefaultNodes;
